+++ CHANGELOG.md
@@ -46,6 +46,7 @@ Upgrading to this release from a v7 release (or later) is seamless and been exte
 - [PR #2519](https://github.com/rqlite/rqlite/pull/2519): Make Raft Config clearer.
 - [PR #2521](https://github.com/rqlite/rqlite/pull/2521): Calculate and check CRCs during reaping.
 - [PR #2524](https://github.com/rqlite/rqlite/pull/2524): Use `AtomicBool` for Snapshot reap-disabled state.
+- [PR #2527](https://github.com/rqlite/rqlite/pull/2527): Read CRC32 values from sidecar files when opening snapshots.
 
 ## v9.4.1 (February 11th 2026)
 ### Implementation changes and bug fixes